+++ init_test.go
@@ -1,0 +1,7 @@
+package httpmux_test
+
+import . "github.com/gbrlsnchs/httpmux"
+
+func init() {
+	SetParamsKey("params")
+}